refactor: Command 모듈의 RabbitMQ 서비스를 Infra 모듈로 이동
[#64]
