+++ command/build.gradle.kts
@@ -8,9 +8,10 @@ plugins {
 dependencies {
     implementation(project(":common"))
     implementation(project(":domain:mysql"))
+    implementation(project(":infrastructure:rabbitmq"))
     implementation(project(":infrastructure:rdbms:jpa"))
 
-    implementation("org.springframework.boot:spring-boot-starter-amqp")
+    implementation("org.springframework:spring-tx:5.2.16.RELEASE")
     implementation("org.springframework.boot:spring-boot-starter-validation")
     implementation("org.springframework.boot:spring-boot-starter-web")
 }
