+++ command/src/main/kotlin/com/hs/application/handler/ProductEventHandler.kt
@@ -5,7 +5,7 @@ import com.hs.application.handler.event.ProductChangeConfirmStatusEvent
 import com.hs.application.handler.event.ProductCreateAndUpdateEvent
 import com.hs.application.handler.event.ProductDecreaseStockQuantityEvent
 import com.hs.application.handler.event.ProductUpdateImageEvent
-import com.hs.infrastructure.rabbitmq.ProductQueuePublisher
+import com.hs.publisher.ProductQueuePublisher
 import com.hs.repository.ProductEventLogRepository
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
