+++ infrastructure/rabbitmq/build.gradle.kts
@@ -1,12 +1,7 @@
 plugins {
-    id("org.springframework.boot")
-
-    kotlin("jvm")
     kotlin("plugin.spring")
-    kotlin("plugin.allopen")
-    kotlin("plugin.jpa")
-    kotlin("kapt")
 }
 
 dependencies {
+    implementation("org.springframework.boot:spring-boot-starter-amqp")
 }
