+++ infrastructure/rabbitmq/src/main/kotlin/com/hs/config/RabbitMqConfig.kt
