+++ infrastructure/rabbitmq/src/main/kotlin/com/hs/event/PublishProductEvent.kt
@@ -1,4 +1,4 @@
-package com.hs.infrastructure.rabbitmq.event
+package com.hs.event
 
 data class PublishProductEvent(
     val productId: Long
