+++ infrastructure/rabbitmq/src/main/kotlin/com/hs/publisher/ProductQueuePublisher.kt
@@ -1,4 +1,4 @@
-package com.hs.infrastructure.rabbitmq
+package com.hs.publisher
 
 interface ProductQueuePublisher {
     fun publish(productId: Long)
